Disable FastFail and switch to Pytest (#39)
* Disable FastFail and switch to Pytest

continue-on-error makes the `master` tests always appear green even if they fail. It doesn't allow `master` and `PyPi` to be tested independently, as seems to have been the intention. The `fail-fast: false` line in strategy should fix that.
Nose hasn't been updated in a few years so we can safely consider it deprecated. Replaced with pytest.

* Forgot pytest-xdist

* Rename test.py files to comply with pytest standard